refactor(auth): rename db file from auth to user
