+++ packages/db/src/features/auth/index.ts
@@ -1,1 +1,1 @@
-export * from "./auth.schema";
+export * from "./user.schema";
